+++ routes/user.js
@@ -4,15 +4,23 @@ const userService = require('../services/userService');
 const activityService = require('../services/activityLogsService');
 const jwtMiddleware = require('../jwt');
 const router = express.Router();
+const adminRol = "ADMIN"
+const permisosInsuficientes = "El usuario no tiene los permisos necesarios para realizar la operacion"
 
 // Get users
 router.get('/', async (req, res) => {
     try {
+        const tokenParsed = jwtMiddleware.verifyAndParseToken(req);
+        const rolTokenValidation = await jwtMiddleware.getRolFromToken(tokenParsed);
+        if(rolTokenValidation!== adminRol){
+            console.log(permisosInsuficientes)
+            throw Error(permisosInsuficientes);
+        }
         const users = await userService.getUsers()
         res.json(users);
     } catch (err) {
-        console.error('Error al obtener los usuarios:', err);
-        res.status(500).json({ error: 'Error al obtener los usuarios' });
+        console.error('Error al obtener los usuarios:', err.toString());
+        res.status(500).json({ error: 'Error al obtener los usuarios.' + err.toString() });
     }
 });
 
@@ -24,7 +32,8 @@ router.get("/:id", async (req, res) => {
     try {
         const tokenParsed = jwtMiddleware.verifyAndParseToken(req);
         const validationToken = await jwtMiddleware.tokenValidationWithId(tokenParsed, userId);
-        if(validationToken){
+        const rolTokenValidation = await jwtMiddleware.getRolFromToken(tokenParsed);
+        if(validationToken && rolTokenValidation!==adminRol){
             throw Error(validationToken);
         }
         const user = await userService.getUser(userId);
@@ -50,9 +59,9 @@ router.get("/:id", async (req, res) => {
 
 //Create user
 router.post('/', async (req, res) => {
-    const { name, lastname, dni, email, password } = req.body;
+    const { name, lastname, dni, email, password, rol } = req.body;
     try {
-        const user = await userService.createUser({ name, lastname, dni, email, password });
+        const user = await userService.createUser({ name, lastname, dni, email, password, rol });
         res.json(user);
     } catch (err) {
         console.error('Error al crear el usuario:', err);
@@ -63,14 +72,24 @@ router.post('/', async (req, res) => {
 //Delete user
 router.delete("/:id",async (req, res) => {
     const userId = req.params.id;
+    //TODO agregar validacion de id nulo, 400
     try {
+        const tokenParsed = jwtMiddleware.verifyAndParseToken(req);
+        const rolTokenValidation = await jwtMiddleware.getRolFromToken(tokenParsed);
+        if(rolTokenValidation!== adminRol){
+            console.log(permisosInsuficientes)
+            res.status(401).json({ error: permisosInsuficientes });
+            return
+        }
+
         const user = await userService.getUser(userId)
         if (user) {
              const deletedUser = await userService.deleteUser(user)
             res.status(200).json({deletedUser})
-        } else {
-            res.status(404).json({ error: 'Usuario no encontrado' });
+            return
         }
+        res.status(404).json({ error: 'Usuario no encontrado' });
+        return
     }  catch (err) {
         console.error('Error al buscar el usuario:', err);
         res.status(500).json({ error: 'Error al buscar el usuario' });
